+++ src/card/cards/coin_flips.rs
@@ -5,3 +5,4 @@ use crate::CardDefinitionId;
 pub const DJINN_TOKEN_5_5_COLORLESS: CardDefinitionId = CardDefinitionId(1481);
 pub const ORCISH_CAPTAIN: CardDefinitionId = CardDefinitionId(1482);
 pub const BOTTLE_OF_SULEIMAN: CardDefinitionId = CardDefinitionId(1483);
+pub const GOBLIN_KITES: CardDefinitionId = CardDefinitionId(1800);